Amplitude and phase functions

+++ inc/sensor_bridge.h
@@ -17,6 +17,6 @@ void interface_write_helper(void);
 // Command line parser 
 void parse_commandline(int argc, char **argv, char *&uart_name, int &baudrate);
 int phase (float, float);
-int amplitude (float, float);
+float amplitude (float, float);
 
 #endif
